// src/app/models/OrderTransaction.ts
export class orderTransaction{
     orderUniqueId: number | undefined;
     memberId: number | undefined;
     orderTotalPrice: number | undefined;
     productUniqueid: number | undefined;
     productQuantity: number | undefined;
     productPrice: number | undefined;
     cardType: String | undefined;
     cardNumber: number | undefined;
     cardExpiry: number | undefined;
     cardCvv: String | undefined;
     cardBrand: String | undefined;
     cardOwnerName: String | undefined;
    
     
}